backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
google-cloud-aiplatform
googlemaps
python-dotenv
